+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 28.3.0
+
+## Bugfixes
+
+### Fix product media gallery cleanup logic
+* Implement entity-specific product media gallery deletion and cleanup logic
+
 # Version 28.2.0
 
 ## Features